Use pre-wrap for run-system
Previously used post-wrap, which didn't do anything in combination with wait as
the wait would never end & trigger the post condition.

I believe this was designed as such for enforcing that the assets are build
before starting the system. I'd recommend this was instead achieved by moving
the run-system task further down the pipeline.