Automatically switch to dev
This wraps rebel manually, as currently rebel does not support arguments.  Upon
failure to load `dev`, the `user` namespace is loaded with a `fixedit!`
function which can be run to continue.

`dev` function is kept in `user` so that this works for normal invocations of
`clj`, without the special rebel.