Replace /public/% with /%
This is to avoid using 'public' in URLs, where public is implied. The
intention behind 'public' was to indicate sub-directories in the
classpath that are meant to be published.

This has also required changes to kick.alpha.

This commit has also fixed browser errors found in Firefox that were
caused by asciidoc source blocks in the documentation not being marked
with a language and causing a runtime issue in highlightjs.